% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{coef}
\alias{coef}
\title{Regression coefficients}
\usage{
coef(x)
}
\arguments{
\item{x}{An object of class linreg contanining a linear regression.}
}
\value{
A numeric vector with the estimated coefficients
}
\description{
\code{plot.linreg} Return estimated coefficients from a linear regression model
}
\details{
This function returns estimated coefficients of a linear regression stored in 
an S3 object of class linreg.
}
