% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg_ols}
\alias{linreg_ols}
\title{Basic OLS Linear Regression.}
\usage{
linreg_ols(data, formula, type)
}
\arguments{
\item{data}{An input data frame}

\item{formula}{A formula object contanining the model.}
}
\value{
An S3 object of class linreg
}
\description{
\code{linreg_ols} returns a S3 object of class linreg
}
\details{
This function performs a basic linear regression model and return an object
for which a number of methods are defined to analyze the regression results.
Estimation is done using Ordinary-Least-Squares method.
}
\seealso{
\url{https://en.wikipedia.org/wiki/Linear_regression}
}
