\name{mynewrpackage-package}
\alias{mynewrpackage-package}
\alias{mynewrpackage}
\docType{package}
\title{
\packageTitle{mynewrpackage}
}
\description{
\packageDescription{mynewrpackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mynewrpackage}
\packageIndices{mynewrpackage}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mynewrpackage}

Maintainer: \packageMaintainer{mynewrpackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
