% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{plot}
\alias{plot}
\title{Residual plots.}
\usage{
plot(x)
}
\arguments{
\item{x}{An object of class linreg contanining a linear regression.}
}
\description{
\code{plot.linreg} Generate residual plots of a linear regression model
}
\details{
This function generate residual plots of a linear regression stored in an
S3 object of class linreg.
}
