% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{pred}
\alias{pred}
\title{Fitted values}
\usage{
pred(x)
}
\arguments{
\item{x}{An object of class linreg contanining a linear regression.}
}
\value{
A numeric vector of fitted values
}
\description{
\code{plot.linreg} Return fitted values from a linear regression model
}
\details{
This function returns the fitted values of a linear regression stored in 
an S3 object of class linreg.
}
