% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{print.linreg}
\alias{print.linreg}
\title{Print coefficients.}
\usage{
\method{print}{linreg}(x)
}
\arguments{
\item{x}{An object of class linreg contanining a linear regression.}
}
\value{
An S3 object of class linreg
}
\description{
\code{print.linreg} Print the coefficients of a linear regression model
}
\details{
This function print the coefficients of a linear regression stored in an
object of class linreg.
}
