% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{resid}
\alias{resid}
\title{Residuals}
\usage{
resid(x)
}
\arguments{
\item{x}{An object of class linreg contanining a linear regression.}
}
\value{
A numeric vector of residuals
}
\description{
\code{plot.linreg} Return vector of residuals from a linear regression model
}
\details{
This function returns a vector of residuals from a linear regression stored in 
an S3 object of class linreg.
}
