% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{summary}
\alias{summary}
\title{Regression summary}
\usage{
summary(x)
}
\arguments{
\item{x}{An object of class linreg contanining a linear regression.}
}
\description{
\code{plot.linreg} Displays summary information of a fitted linear regression model
}
\details{
This function displays the estimated coefficients and associated t-values and p-values
of a linear regression stored in an S3 object of class linreg. Also the estimated error
variance of the model and the degrees of freedom are shown.
}
