% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Basic Linear Regression.}
\usage{
linreg(data, formula)
}
\arguments{
\item{data}{An input data frame}

\item{formula}{A formula object contanining the model.}
}
\value{
An S3 object of class linreg
}
\description{
\code{linreg} returns a S3 object of class linreg
}
\details{
This function performs a basic linear regression model and return an object
for which a number of methods are defined to analyze the regression results.
}
\seealso{
\url{https://en.wikipedia.org/wiki/Linear_regression}
}
