\name{mynewrpackage-package}
\alias{mynewrpackage-package}
\alias{mynewrpackage}
\docType{package}
\title{
\packageTitle{mynewrpackage}
}
\description{
\packageDescription{mynewrpackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mynewrpackage}
\packageIndices{mynewrpackage}
Esta es una descripcion sencilla
}
\author{
\packageAuthor{mynewrpackage}

Maintainer: \packageMaintainer{mynewrpackage}
}
\references{
No tengo ninguna referencia
}
